just-another-deep-research-agent/pyproject.toml
[project]
name = "just-another-deep-research-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Priyanshu Yashwant Deshmukh",email = "priyanshudeshmukh4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-researcher (>=0.0.10,<0.0.11)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "just_another_deep_research_agent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
